feat: add requirements and return specific error on unsupported environment
