+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/cloudwego/dynamicgo
 
-go 1.15
+go 1.16
 
 require (
 	github.com/apache/thrift v0.13.0
@@ -23,5 +23,4 @@ require (
 	golang.org/x/arch v0.2.0 // indirect
 	google.golang.org/genproto v0.0.0-20230222225845-10f96fb3dbec // indirect
 	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
-	gopkg.in/yaml.v2 v2.4.0 // indirect
 )
